+++ CHANGELOG.md
@@ -5,14 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+- Add custom ancestor selector
+
 ### Changed
 - Changed demo
 - Updated dev todo's
 
 
 ## [0.1.0] - 2019-10-06
-### Added
 
+### Added
 - Validator Engine class with all the validation logic, options handling and the `notempty`/required validator: `mightyFormValidator.js`
 - Default validators, placed in: `validators/`
   - E-mail validator (`validators/email.js`)